Move formatDateToLocal to utils file
