+++ dashboard/15-final/app/lib/utils.ts
@@ -1,10 +1,24 @@
-import { Invoice, Revenue, Customer, LatestInvoice } from './definitions';
+import { Invoice, Revenue, Customer } from './definitions';
 
 export const formatCurrency = (amount: number) => {
   return (amount / 100).toLocaleString('en-US', {
     style: 'currency',
     currency: 'USD',
   });
+};
+
+export const formatDateToLocal = (
+  dateStr: string,
+  locale: string = 'en-US',
+) => {
+  const date = new Date(dateStr);
+  const options: Intl.DateTimeFormatOptions = {
+    day: 'numeric',
+    month: 'short',
+    year: 'numeric',
+  };
+  const formatter = new Intl.DateTimeFormat(locale, options);
+  return formatter.format(date);
 };
 
 export function findLatestInvoices(invoices: Invoice[], customers: Customer[]) {
