+++ dashboard/15-final/app/ui/invoices/table.tsx
@@ -33,16 +33,7 @@ function renderInvoiceStatus(status: string) {
   }
 }
 
-function formatDateToLocal(dateStr: string, locale: string = 'en-US') {
-  const date = new Date(dateStr);
-  const options: Intl.DateTimeFormatOptions = {
-    day: 'numeric',
-    month: 'short',
-    year: 'numeric',
-  };
-  const formatter = new Intl.DateTimeFormat(locale, options);
-  return formatter.format(date);
-}
+import { formatDateToLocal, formatCurrency } from '@/app/lib/utils';
 
 export default async function InvoicesTable({
   searchParams,
